Add Support for Debugging Binaries
